+++ R/defaultControl.R
@@ -40,7 +40,7 @@
 #'       \item{\code{unlimitedSoilWater [= FALSE]}: Boolean flag to indicate the simulation of plant transpiration assuming that soil water is always at field capacity.}
 #'       \item{\code{unfoldingDD [= 300]}: Degree-days for complete leaf unfolding after budburst has occurred.}
 #'       \item{\code{interceptionMode [= "Gash1995"]}: Infiltration model, either "Gash1995" or "Liu2001".}
-#'       \item{\code{infiltrationMode [= "Green-Ampt"]}: Infiltration model, either "Green-Ampt" or "Boughton".}
+#'       \item{\code{infiltrationMode [= "GreenAmpt1911"]}: Infiltration model, either "GreenAmpt1911" or "Boughton1989".}
 #'       \item{\code{rhizosphereOverlap [= "total"]}: A string indicating the degree of rhizosphere spatial overlap between plant cohorts:
 #'           \itemize{
 #'             \item{"none" - no overlap (independent water pools).}
@@ -195,7 +195,7 @@ defaultControl<-function(transpirationMode = "Granier") {
     bareSoilEvaporation = TRUE,
     unlimitedSoilWater = FALSE,
     interceptionMode = "Gash1995",
-    infiltrationMode = "Green-Ampt",
+    infiltrationMode = "GreenAmpt1911",
     rhizosphereOverlap = "total",
     unfoldingDD = 300,
     verticalLayerSize = 100,